pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "arabic-reshaper>=3.0.0",
    "bcrypt>=5.0.0",
    "beautifulsoup4>=4.14.3",
    "cairosvg>=2.8.2",
    "extra-streamlit-components>=0.1.81",
    "fpdf>=1.7.2",
    "fpdf2>=2.8.5",
    "google-genai>=1.55.0",
    "jinja2>=3.1.6",
    "lxml>=6.0.2",
    "openpyxl>=3.1.5",
    "pillow>=12.0.0",
    "playwright>=1.57.0",
    "psycopg2-binary>=2.9.11",
    "python-bidi>=0.6.7",
    "requests>=2.32.5",
    "resend>=2.19.0",
    "sift-stack-py>=0.9.4",
    "sqlalchemy>=2.0.45",
    "streamlit>=1.52.1",
    "streamlit-drawable-canvas>=0.9.3",
    "streamlit-paste-button>=0.1.2",
    "weasyprint>=67.0",
]
